Rename `text-style-bar.*` to `text-toolbar.*`
