+++ src/components/tools/text-toolbar.sass
@@ -7,9 +7,6 @@
   justify-content: space-around
   align-items: center
   margin: 4px 2px 2px 2px
-  float: left
-  top: 86px
-  left: 0
   width: 232px
   height: 29px
